Повышенная сложность

+++ src/Employee.java
@@ -52,7 +52,7 @@ public class Employee {
         this.department = department;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
